% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clutches_to_sample.R
\name{clutches_to_sample}
\alias{clutches_to_sample}
\title{clutches_to_sample}
\usage{
clutches_to_sample(
  n_sims = 10000,
  pop_size = 100,
  sample_size = 32,
  paternal_contribution_mode = "random",
  Fprob,
  Mprob,
  clutches_mu = 4.95,
  clutches_sd = 2.09,
  prop_correct
)
}
\arguments{
\item{n_sims}{integer value, the number of simulations to run. Default value
is 10000.}

\item{pop_size}{integer value, the population size of all breeding adults.
Default value is 100.}

\item{sample_size}{integer value, the sample size to collect. Default value
is 32.}

\item{paternal_contribution_mode}{a character value defining the distribution
of father contributions to fertilizing a single clutch. Potential values
include random', 'exponential', dominant50', 'dominant70',
'dominant90', 'mixed_dominant'). Default value is 'random'.}

\item{Fprob}{a numeric vector, the probabilities of eggs from 1 mother being
fertilized by 1-max_fathers fathers.}

\item{Mprob}{a numeric vector, the probabilities of fathers fertilizing eggs
from 1+ mothers}

\item{clutches_mu}{a  numeric value, the mean number of clutches a mother
lays in one nesting season. Default value is 4.95.}

\item{clutches_sd}{a numeric value, the standard deviation of the number of
clutches a mother lays in one nesting season. Default value is 2.09.}

\item{proportion_correct_all}{a data frame with columns "Paternal
Contribution Mode", "Fathers" (number of contributing fathers),
"Sample Size" (1 - 96), "Proportion Correct" (how many simulations
correctly identified all fathers), and "Marginal" (the marginal
paternal contributions of the last father).}
}
\value{
returns a data frame with the proportion of simulations where all
fathers are identified given the average and standard deviation of the
numbers of hatchlings in a clutch, the maximum potential number of
fathers, the number of simulations, the number of hatchlings to sample
from clutches, and the paternal contribution mode.
}
\description{
\code{clutches_to_sample} Samples clutches across a whole breeding season to
determine if all of the fathers that contributed were identified.
}
\examples{
proportion_correct_all <- hatchlings_to_sample(
                                       hatchlings_mu = 100.58,
                                       hatchlings_sd = 22.61,
                                       max_fathers = 5,
                                       n_sims = 1000,
                                       sample_sizes = c(32, 96),
                                       paternal_contribution_mode = 'random')

clutches_to_sample(n_sims = 1000,
                pop_size = 100,
                sample_size = 32,
                paternal_contribution_mode = 'random',
                Fprob = c(0.463, 0.318, 0.157, 0.034, 0.028),
                Mprob = c(1),
                clutches_mu = 4.95,
                clutches_sd = 2.09,
                prop_correct = proportion_correct_all)
}
