% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatchlings_to_sample.R
\name{hatchlings_to_sample}
\alias{hatchlings_to_sample}
\title{hatchlings_to_sample}
\usage{
hatchlings_to_sample(
  hatchlings_mu = 100.58,
  hatchlings_sd = 22.61,
  max_fathers = 5,
  n_sims = 10000,
  sample_sizes = c(32, 96),
  paternal_contribution_mode = "random"
)
}
\arguments{
\item{hatchlings_mu}{numeric value, the mean number of hatchlings produced in
a clutch. Default value is 100.58.}

\item{hatchlings_sd}{numeric value, the standard deviation of the number of
hatchlings produced in a clutch. Default value is 22.61.}

\item{max_fathers}{integer value, the maximum number of fathers that mothers
can mate with. Default value is 5.}

\item{n_sims}{integer value, the number of simulations to run. Default value
is 10000.}

\item{sample_sizes}{vector of integer values, the sample size(s) to collect.
Default value is c(32, 96).}

\item{paternal_contribution_mode}{a character value defining the distribution
of paternal contributions to a single clutch. Potential values
include random', 'exponential', dominant50', 'dominant70',
'dominant90', 'mixed_dominant'). Default value is 'random'.}
}
\value{
creates and saves figures to plot confidence of identifying all
fathers given different numbers of actual fathers and sample sizes.
Creates and saves table of confidences for specific sample sizes.
}
\description{
\code{hatchlings_to_sample}  samples hatchlings from clutches to determine
the confidence of identifying all of the fathers that contributed for
populations that exhibit multiple paternity.
}
\examples{
hatchlings_to_sample(hatchlings_mu = 100.58,
                     hatchlings_sd = 22.61,
                     max_fathers = 5,
                     n_sims = 100,
                     sample_sizes = c(32, 96),
                     paternal_contribution_mode = 'random')

}
