% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_id_fathers.R
\name{probability_id_fathers}
\alias{probability_id_fathers}
\title{probability_id_fathers}
\usage{
probability_id_fathers(
  hatchlings_mu = 100.58,
  hatchlings_sd = 22.61,
  max_fathers = 5,
  n_sims = 10000,
  sample_sizes = c(32, 96),
  paternal_contribution_modes = c("random", "exponential", "dominant50", "dominant70",
    "dominant90", "mixed_dominant"),
  min_clutch_size = 10
)
}
\arguments{
\item{hatchlings_mu}{numeric value, the mean number of hatchlings produced in
a clutch. Default value is 100.58.}

\item{hatchlings_sd}{numeric value, the standard deviation of the number of
hatchlings produced in a clutch. Default value is 22.61.}

\item{max_fathers}{integer value, the maximum number of fathers that can
fertilize the eggs from a single mother. Default value is 5.}

\item{n_sims}{integer value, the number of simulations to run. Default value
is 10000.}

\item{sample_sizes}{vector of integer values, the sample size(s) to collect.
Default value is c(32, 96).}

\item{paternal_contribution_modes}{vector of character values, the different
paternal contribution modes to analyse. Default value is c('random',
'exponential', 'dominant50', 'dominant70', 'dominant90', 'mixed_dominant').}

\item{min_clutch_size}{integer value, the number of hatchlings to be assigned
to a clutch when 0 was returned from the normal distribution based on
hatchlings_mu and hatchlings_sd. Default value is 10.}
}
\value{
saves object with probabilities of identifying the number of actual
fathers given paternal contribution mode, sample size, and number of
fathers
}
\description{
\code{probability_id_fathers} saves a data frame recording the probabilities
of detecting 1 - max_fathers fathers given a specific number of potential
fathers, sample size, and paternal contribution mode.
}
\examples{
probability_id_fathers(hatchlings_mu = 100.58,
                       hatchlings_sd = 22.61,
                       max_fathers = 5,
                       n_sims = 100,
                       sample_sizes = c(32, 96),
                       paternal_contribution_modes = c('random',
                                                       'exponential',
                                                       'dominant50',
                                                       'dominant70',
                                                       'dominant90',
                                                       'mixed_dominant'),
                       min_clutch_size = 10)
}
