% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_probabilities.R
\name{conditional_probabilities}
\alias{conditional_probabilities}
\title{conditional_probabilities}
\usage{
conditional_probabilities(probabilities)
}
\arguments{
\item{probabilities}{a data frame that has columns for

character values of the different paternal contribution modes to analyse.
Default values are c('random', 'exponential', 'dominant50', 'dominant70',
'dominant90', 'mixed_dominant').

integer numbers of actual fathers for a clutch. Default values are from
1 to 5.

integer numbers of hatchlings to sample from a clutch. Default values are
32 and 96.

integer numbers of fathers that were observed.

the numeric probability of the number of fathers observed given the number
of actual fathers. Values must be between 0 and 1, inclusive.}
}
\value{
two data frames of conditional probabilities for the number of
actual fathers given the number of fathers identified. The first data frame
has columns for paternal contribution modes, sample sizes, the number of
fathers observed, the number of actual fathers, and a probability column that
is the new conditional probability. The second data frame is wider so that
the conditional probability values are displayed by rows of paternal
contribution mode and observed numbers of fathers with columns for actual
numbers of fathers.
}
\description{
conditional_probabilities
}
\examples{
output <- probability_id_fathers(hatchlings_mu = 100.58,
                                 hatchlings_sd = 22.61,
                                 max_fathers = 5,
                                 n_sims = 10000,
                                 sample_sizes = c(32, 96),
                                 paternal_contribution_modes =
                                    c('random', 'exponential',
                                      'dominant50', 'dominant70',
                                      'dominant90', 'mixed_dominant'),
                                 min_clutch_size = 10)

probabilities <- output[[1]]

conditional_probabilities(probabilities)
}
